Modified jstree default context menu behavior. Added bookmark icon. Got create/rename/delete functionality working.

+++ scripts/jstree/jstree.js
@@ -5787,7 +5787,7 @@
 				"create" : {
 					"separator_before"	: false,
 					"separator_after"	: true,
-					"_disabled"			: false, //(this.check("create_node", data.reference, {}, "last")),
+					"_disabled"			: o.type == 'file', //(this.check("create_node", data.reference, {}, "last")),
 					"label"				: "Create",
 					"action"			: function (data) {
 						var inst = $.jstree.reference(data.reference),
@@ -5800,7 +5800,7 @@
 				"rename" : {
 					"separator_before"	: false,
 					"separator_after"	: false,
-					"_disabled"			: false, //(this.check("rename_node", data.reference, this.get_parent(data.reference), "")),
+					"_disabled"			: o.type == 'file', //(this.check("rename_node", data.reference, this.get_parent(data.reference), "")),
 					"label"				: "Rename",
 					/*!
 					"shortcut"			: 113,
@@ -5817,11 +5817,16 @@
 					"separator_before"	: false,
 					"icon"				: false,
 					"separator_after"	: false,
-					"_disabled"			: false, //(this.check("delete_node", data.reference, this.get_parent(data.reference), "")),
+					"_disabled"			: o.id == 'node_0', //(this.check("delete_node", data.reference, this.get_parent(data.reference), "")),
 					"label"				: "Delete",
 					"action"			: function (data) {
 						var inst = $.jstree.reference(data.reference),
 							obj = inst.get_node(data.reference);
+                        if (obj.type == 'default' && obj.children.length > 0) {
+                            if (!(confirm('Are you sure you want to delete this folder and all subfolders and bookmarks within it?'))) {
+                                return true;
+                            }
+                        }
 						if(inst.is_selected(obj)) {
 							inst.delete_node(inst.get_selected());
 						}
@@ -5830,60 +5835,60 @@
 						}
 					}
 				},
-				"ccp" : {
-					"separator_before"	: true,
-					"icon"				: false,
-					"separator_after"	: false,
-					"label"				: "Edit",
-					"action"			: false,
-					"submenu" : {
-						"cut" : {
-							"separator_before"	: false,
-							"separator_after"	: false,
-							"label"				: "Cut",
-							"action"			: function (data) {
-								var inst = $.jstree.reference(data.reference),
-									obj = inst.get_node(data.reference);
-								if(inst.is_selected(obj)) {
-									inst.cut(inst.get_top_selected());
-								}
-								else {
-									inst.cut(obj);
-								}
-							}
-						},
-						"copy" : {
-							"separator_before"	: false,
-							"icon"				: false,
-							"separator_after"	: false,
-							"label"				: "Copy",
-							"action"			: function (data) {
-								var inst = $.jstree.reference(data.reference),
-									obj = inst.get_node(data.reference);
-								if(inst.is_selected(obj)) {
-									inst.copy(inst.get_top_selected());
-								}
-								else {
-									inst.copy(obj);
-								}
-							}
-						},
-						"paste" : {
-							"separator_before"	: false,
-							"icon"				: false,
-							"_disabled"			: function (data) {
-								return !$.jstree.reference(data.reference).can_paste();
-							},
-							"separator_after"	: false,
-							"label"				: "Paste",
-							"action"			: function (data) {
-								var inst = $.jstree.reference(data.reference),
-									obj = inst.get_node(data.reference);
-								inst.paste(obj);
-							}
-						}
-					}
-				}
+				// "ccp" : {
+				// 	"separator_before"	: true,
+				// 	"icon"				: false,
+				// 	"separator_after"	: false,
+				// 	"label"				: "Edit",
+				// 	"action"			: false,
+				// 	"submenu" : {
+				// 		"cut" : {
+				// 			"separator_before"	: false,
+				// 			"separator_after"	: false,
+				// 			"label"				: "Cut",
+				// 			"action"			: function (data) {
+				// 				var inst = $.jstree.reference(data.reference),
+				// 					obj = inst.get_node(data.reference);
+				// 				if(inst.is_selected(obj)) {
+				// 					inst.cut(inst.get_top_selected());
+				// 				}
+				// 				else {
+				// 					inst.cut(obj);
+				// 				}
+				// 			}
+				// 		},
+				// 		"copy" : {
+				// 			"separator_before"	: false,
+				// 			"icon"				: false,
+				// 			"separator_after"	: false,
+				// 			"label"				: "Copy",
+				// 			"action"			: function (data) {
+				// 				var inst = $.jstree.reference(data.reference),
+				// 					obj = inst.get_node(data.reference);
+				// 				if(inst.is_selected(obj)) {
+				// 					inst.copy(inst.get_top_selected());
+				// 				}
+				// 				else {
+				// 					inst.copy(obj);
+				// 				}
+				// 			}
+				// 		},
+				// 		"paste" : {
+				// 			"separator_before"	: false,
+				// 			"icon"				: false,
+				// 			"_disabled"			: function (data) {
+				// 				return !$.jstree.reference(data.reference).can_paste();
+				// 			},
+				// 			"separator_after"	: false,
+				// 			"label"				: "Paste",
+				// 			"action"			: function (data) {
+				// 				var inst = $.jstree.reference(data.reference),
+				// 					obj = inst.get_node(data.reference);
+				// 				inst.paste(obj);
+				// 			}
+				// 		}
+				// 	}
+				// }
 			};
 		}
 	};
